Fix for bug #566821 and cleanups to packaging process. With these changes, it builds under both testing and karmic correctly. We'll check the other systems ASAP.
*: adding a AC_CONFIG_AUX_DIR([.]) to configure.in seems to eliminate
the need for a duplicate libtoolize (autoreconf also calls libtoolize,
making it a dupe) _and_ fixes the problem described in bug #566821 where
ltmain.sh isn't being created on older systems.

*: remove the calls to fakeroot. From the fakeroot man page:

  "Also, note that it's best not to do the building of the binaries themselves under  fakeroot.  Especially
  configure  and friends don't like it when the system suddenly behaves differently from what they expect.
  (or, they randomly unset some environment variables, some of which fakeroot needs)."

So inserting fakeroot where we were in the packaging process was confusing things. You _only_ need it
during a specific portion of the build where you do the "make install" phase, and according to the
dpkg-buildpackage man page:

 "5. It calls debian/rules build followed by fakeroot debian/rules  binary-target  (unless  a
  source-only build  has  been requested with -S). Note that binary-target is either binary (default case,
  or if -b is specified) or binary-arch (if -B is specified) or binary-indep (if -A is specified)."

In other words, it's being correctly handled for us, so we don't need to explicitly use it.