Fix problem related to premature end-of-data while reading datasets.
Found myself a test sample where my existing tables apparently have a larger
datawidth for one or more descriptors. This leads to running out of bits when we
hit the fifth data subset. The bufr_decode_message(), however, was running out
of bits and just dropping the last data subset while, in fact, ALL the data
subsets were essentially garbage.

This patch alters the behaviour such that if we run out of bits during the
decoding process, the entire dataset is flagged as invalid.

We might want to explore what to do if bits are leftover since that
suggests a datawidth mismatch in the opposite direction. However, IIRC the BUFR
spec doesn't restrict the amount of extra padding in sections, so it's probably
not feasible.

Vanh, I think you'll want to review this patch...