Fixes bug Bug #826868
Make sure when we output a string that it's space filled (or truncated) to the
descriptor datawidth. This means (a) we're not trusting the user to get it
right, and (b) the user can think of strings in a C context without having to
keep track of data width.

When comparing BUFR strings for output compression, do so using BUFR semantics
rather than a straight strncmp().

Fix some documentation where it might not be clear how strings get handled.