Fix for bug #1193035.
We define a "default" floating point valtype VALTYPE_FLTDEFAULT which
is either VALTYPE_FLT64 (default) or VALTYPE_FLT32. This is used where
we're not dealing with a specific IEEE FP length. For precision reasons,
we prefer the 64-bit but the option to revert to 32 is there for numerical
compatibility with prior versions.

The immediate impact is that BUFR value/scale/ref decode (and probably
encode) calculations are done using that precision and test_zero passes (what
goes in is what comes out).

This changes regression output considerably in places. Most of it is
cosmetic, as a side-effect of trailing zeros in vprintf() formatting. The
most interesting difference is things like:

-   002153 {R=0} VALUE=24876800802816.0
-   002154 {R=0} VALUE=1883700068352.0
+   002153 {R=0} VALUE=24876800000000
+   002154 {R=0} VALUE=1883700000000

002153 has a scale of -8, which means the lower 8 digits _should_ be zero.
Which points to a separate bug in the 32-bit floating point logic.